Remove MeTTaPy implementation
